# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A part of microservice infrastructure, who responsible for reCAPTCHA v2/v3 validation",
        "title": "reCAPTCHA Service",
        "contact": {
            "name": "Artem Kostenko",
            "url": "https://github.com/aerosystems"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "gw.verifire.com/recaptcha",
    "basePath": "/",
    "paths": {
        "/v1/validate": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-gateway-special"
                ],
                "summary": "validate reCAPTCHA v3 token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reCAPTCHA v3 token",
                        "name": "X-Recaptcha-V3-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "reCAPTCHA v2 token",
                        "name": "X-Recaptcha-V2-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
        "/v1/validate/v3": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-gateway-special"
                ],
                "summary": "validate reCAPTCHA v3 token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reCAPTCHA v3 token",
                        "name": "X-Recaptcha-V3-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}